+++ src/cmd/aliases.go
@@ -5,6 +5,7 @@ import (
 	"strings"
 
 	"github.com/git-town/git-town/v7/src/config"
+	"github.com/git-town/git-town/v7/src/flags"
 	"github.com/git-town/git-town/v7/src/git"
 	"github.com/spf13/cobra"
 )
@@ -20,7 +21,7 @@ Does not overwrite existing aliases.
 This can conflict with other tools that also define Git aliases.`
 
 func aliasesCommand() *cobra.Command {
-	addDebugFlag, readDebugFlag := debugFlag()
+	addDebugFlag, readDebugFlag := flags.Debug()
 	cmd := cobra.Command{
 		Use:     "aliases (add | remove)",
 		GroupID: "setup",
