+++ src/cmd/config.go
@@ -5,6 +5,7 @@ import (
 	"strings"
 
 	"github.com/git-town/git-town/v7/src/cli"
+	"github.com/git-town/git-town/v7/src/flags"
 	"github.com/git-town/git-town/v7/src/runstate"
 	"github.com/spf13/cobra"
 )
@@ -12,7 +13,7 @@ import (
 const configDesc = "Displays your Git Town configuration"
 
 func configCmd() *cobra.Command {
-	addDebugFlag, readDebugFlag := debugFlag()
+	addDebugFlag, readDebugFlag := flags.Debug()
 	configCmd := cobra.Command{
 		Use:     "config",
 		GroupID: "setup",
